Move FootpathEntry
